Изменил дизайн order_body_item.xml, добавил еще один view_type, теперь если цена была изменена, то эллемент становится темным. Изменил тип колонки цена и сумма с Double на Int. Теперь при создании заказа автоматически открывается окно с его телом

+++ app/src/main/java/com/example/orderlite/data/orderRecord/OrderRecordDbModel.kt
@@ -27,7 +27,7 @@ data class OrderRecordDbModel(
     @ColumnInfo(name = "unit_id")
     val unitId:Int,
     @ColumnInfo(name = "price")
-    val price:Double,
+    val price:Int,
     @ColumnInfo(name = "orderRecordAmount")
     val amount:Double
 )
